fix í:파사ë Template command단 퍼사드 패턴 ㅈ적용
