--- src/main/java/com/ideality/coreflow/attachment/command/application/service/AttachmentService.java
@@ -1,15 +1,0 @@
-package com.ideality.coreflow.attachment.command.application.service;
-
-import org.springframework.stereotype.Service;
-
-import com.ideality.coreflow.attachment.command.domain.repository.AttachmentRepository;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class AttachmentService {
-
-	private final AttachmentRepository attachmentRepository;
-
-}
