+++ src/main/java/com/ideality/coreflow/template/query/service/TemplateQueryFacadeService.java
@@ -1,0 +1,13 @@
+package com.ideality.coreflow.template.query.service;
+
+import org.springframework.stereotype.Service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+public class TemplateQueryFacadeService {
+
+}
